Fixed the ??? bug in hovertool
Both IP and NIP hovertool is added a 'renderers' property that tells which scatter series (IP or NIP) it is to be applied to. That way, the hovering for both IP and NIP would not attempt to pop-up '???' values as it doesn't recognize the other set of point.

The print statements in main is also removed as it causes asynchronous bug for CSV processing.